requirements.txt
# FastAPI y servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Base de datos
supabase>=2.0.0
python-dotenv>=1.0.0

# Data Science y ML
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
scipy==1.11.4

# Deep Learning
torch==2.1.2
torchvision==0.16.2

# Utilidades
joblib==1.3.2
openpyxl==3.1.2
xlrd==2.0.1
python-magic==0.4.27
python-magic-bin==0.4.14;platform_system=="Windows"

# Gráficos y visualización
matplotlib==3.8.2
seaborn==0.13.0

# Logging y monitoreo
colorlog==6.8.0

